@@ -8,14 +8,14 @@
     "build": "webpack"
   },
   "keywords": [],
-  "author": "Lothric",
+  "author": "",
   "license": "ISC",
+  "devDependencies": {
+    "webpack": "^5.65.0",
+    "webpack-cli": "^4.9.1"
+  },
   "dependencies": {
     "jquery": "^3.6.0",
     "lodash": "^4.17.21"
-  },
-  "devDependencies": {
-    "webpack": "^5.40.0",
-    "webpack-cli": "^4.7.2"
   }
 }
