@@ -1,25 +1,25 @@
 {
-  "name": "task_2",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack"
-  },
-  "keywords": [],
-  "author": "UWABOR COLLINS",
-  "license": "UWABOR",
-  "dependencies": {
-    "jquery": "^3.6.0",
-    "lodash": "^4.17.21"
-  },
-  "devDependencies": {
-    "css-loader": "^5.2.6",
-    "file-loader": "^6.2.0",
-    "image-webpack-loader": "^7.0.1",
-    "style-loader": "^3.0.0",
-    "webpack": "^5.40.0",
-    "webpack-cli": "^4.7.2"
-  }
+	"name": "task_1",
+	"version": "1.0.0",
+	"description": "",
+	"main": "index.js",
+	"scripts": {
+		"test": "echo \"Error: no test specified\" && exit 1",
+		"build": "webpack"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"devDependencies": {
+		"css-loader": "^6.5.1",
+		"file-loader": "^6.2.0",
+		"image-webpack-loader": "^8.0.1",
+		"style-loader": "^3.3.1",
+		"webpack": "^5.65.0",
+		"webpack-cli": "^4.9.1"
+	},
+	"dependencies": {
+		"jquery": "^3.6.0",
+		"lodash": "^4.17.21"
+	}
 }
