@@ -1,29 +1,29 @@
 {
-  "name": "task_1",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack",
-    "start-dev": "webpack-dev-server --open"
-  },
-  "keywords": [],
-  "author": "Lothric",
-  "license": "ISC",
-  "dependencies": {
-    "jquery": "^3.6.0",
-    "lodash": "^4.17.21"
-  },
-  "devDependencies": {
-    "clean-webpack-plugin": "^4.0.0-alpha.0",
-    "css-loader": "^5.2.6",
-    "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.3.2",
-    "image-webpack-loader": "^7.0.1",
-    "style-loader": "^3.0.0",
-    "webpack": "^5.40.0",
-    "webpack-cli": "^4.7.2",
-    "webpack-dev-server": "^3.11.2"
-  }
+	"name": "task_1",
+	"version": "1.0.0",
+	"description": "",
+	"main": "index.js",
+	"scripts": {
+		"test": "echo \"Error: no test specified\" && exit 1",
+		"build": "webpack",
+		"start-dev": "webpack-dev-server --open"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"devDependencies": {
+		"clean-webpack-plugin": "^4.0.0",
+		"css-loader": "^6.5.1",
+		"file-loader": "^6.2.0",
+		"html-webpack-plugin": "^5.5.0",
+		"image-webpack-loader": "^8.0.1",
+		"style-loader": "^3.3.1",
+		"webpack": "^5.65.0",
+		"webpack-cli": "^4.9.1",
+		"webpack-dev-server": "^4.6.0"
+	},
+	"dependencies": {
+		"jquery": "^3.6.0",
+		"lodash": "^4.17.21"
+	}
 }
