@@ -1,6 +1,5 @@
-"use strict";
-const $ = require("jquery");
-import "./footer.css";
+const $ = require('jquery');
+import './footer.css';
 
-$("body").append("<footer></footer>");
-$("footer").append("<p>Copyright - Holberton School</p>");
+$('body').append('<footer></footer>');
+$('footer').append('<p>Copyright - Holberton School</p>');
