@@ -1,8 +1,7 @@
-"use strict";
-const $ = require("jquery");
-import "./header.css";
+import './header.css';
+const $ = require('jquery');
 
-$("body").append("<header></header>");
-$("header").append('<div id="logo"></div>', "<h1>Holberton Dashboard</h1>");
+$('body').append('<header></header>');
+$('header').append('<div id="logo"></div>', '<h1>Holberton Dashboard</h1>');
 
-console.log("Init header");
+console.log('Init header');
