@@ -1,18 +1,4 @@
+export const getFullYear = () => new Date().getFullYear();
 
-const getFullYear = () => {
-	let current_year = new Date().getFullYear();
-	return current_year;
-}
-
-const getFooterCopy = (isIndex) => {
-	if (isIndex) {
-		return "Holberton School";
-	} else {
-		return "Holberton School main dashboard";
-	}
-}
-
-module.exports = {
-	getFooterCopy,
-	getFullYear,
-};
+export const getFooterCopy = (isIndex) =>
+	isIndex ? 'Holberton School' : 'Holberton School main dashboard';
