@@ -1,19 +1,18 @@
-'use strict';
-import $ from 'jquery';
-import _ from 'lodash';
+const $ = require('jquery');
+const _ = require('lodash');
 
 $('body').append('<p>Holberton Dashboard</p>');
 $('body').append('<p>Dashboard data for the students</p>');
 $('body').append('<button>Click here to get started</button>');
-$('body').append('<p id="count"></p>');
+$('body').append("<p id='count'></p>");
 $('body').append('<p>Copyright - Holberton School</p>');
 
 const updateCounter = () => {
-  let times = $('#count').html() || 0;
-  $('button').on('click', () => {
-    times++;
-    $('#count').html(`${times} clicks on the button`);
-  });
+	let times = $('#count').html() || 0;
+	$('button').on('click', () => {
+		times++;
+		$('#count').html(`${times} clicks on the button`);
+	});
 };
 
 _.debounce(updateCounter, 500);
