@@ -1,10 +1,12 @@
 const path = require('path');
 
 module.exports = {
-  mode: 'production',
-  entry: './js/dashboard_main.js',
-  output: {
-    path: path.resolve(__dirname, 'public'),
-    filename: 'bundle.js',
-  },
+	mode: 'production',
+	entry: {
+		main: path.resolve(__dirname, './js/dashboard_main.js'),
+	},
+	output: {
+		path: path.resolve(__dirname, 'public'),
+		filename: 'bundle.js',
+	},
 };
