@@ -1,18 +1,20 @@
-
-import $ from 'jquery';
-import _ from 'lodash';
+import '../css/main.css';
+const $ = require('jquery');
+const _ = require('lodash');
 
 $('body').append('<p>Holberton Dashboard</p>');
 $('body').append('<p>Dashboard data for the students</p>');
 $('body').append('<button>Click here to get started</button>');
-$('body').append('<p id="count"></p>');
+$('body').append("<p id='count'></p>");
 $('body').append('<p>Copyright - Holberton School</p>');
 
-let conunt = 0;
-
-function updateCounter() {
-  count++;
-  $("#count").html(`${count} clicks on the button`);
+const updateCounter = () => {
+	let times = $('#count').html() || 0;
+	$('button').on('click', () => {
+		times++;
+		$('#count').html(`${times} clicks on the button`);
+	});
 };
 
-$('button').on('click', _.debounce(updateCounter, 500));
+_.debounce(updateCounter, 500);
+updateCounter();
