@@ -1,16 +1,17 @@
-import $ from 'jquery';
-import _ from 'lodash';
 import './body.css';
+const $ = require('jquery');
+const _ = require('lodash');
 
-let count = 0;
-
-$('body').append('<p>Dashboard data for the students</p>');
 $('body').append('<button>Click here to get started</button>');
-$('body').append('<p id="count"></p>');
+$('body').append('<p id="count"><p>');
 
-$('button').on('click', _.debounce(updateCounter, 500));
+const updateCounter = () => {
+	let times = $('#count').html() || 0;
+	$('button').on('click', () => {
+		times++;
+		$('#count').html(`${times} clicks on the button`);
+	});
+};
 
-function updateCounter() {
-  count = count + 1;
-  $('#count').text(count + ' clicks on the button');
-}
+_.debounce(updateCounter, 500);
+updateCounter();
