@@ -1,13 +1,12 @@
-
 import React from 'react';
 import './Notifications.css';
 
-export default function Notifications() {
+const Notifications = () => {
 	return (
-		<div className="Notifications">
-			<p>
-				Here is the list of notifications
-			</p>
+		<div className='Notifications'>
+			<p>Here is the list of notifications</p>
 		</div>
 	);
 };
+
+export default Notifications;
