@@ -1,30 +1,35 @@
-import logo from './logo.jpg';
+import logo from './holberton-logo.jpg';
 import './App.css';
-import { getFooterCopy, getFullYear } from './utils';
+import { getFullYear, getFooterCopy } from './utils';
 
 function App() {
-  return (
-    <div className="App">
-        <div className="App-header">
-        <img src={logo} alt="logo" />
-        <h1>School dashboard</h1>
-      </div>
-      <div className="App-body">
-        <p>
-          Login to access the full dashboard
-        </p>
-		<br></br>
-		<label htmlFor="email">  Email: </label>
-        <input type="email" id="email" name="email" />
-        <label htmlFor="password">  Password: </label>
-        <input type="password" id="password" name="password" />
-        <button>  OK</button>
-      </div>
-      <div className="App-footer">
-	  <p>Copyright {getFullYear()} - {getFooterCopy()}</p>
-      </div>
-    </div>
-  );
+	return (
+		<div className='App'>
+			<div className='App-header'>
+				<img src={logo} alt='Holberton' />
+				<h1>School dashboard</h1>
+			</div>
+			<div className='App-body'>
+				<p>Login to access the full dashboard</p>
+				<section className='form-inputs'>
+					<section className='input'>
+						<label htmlFor='email'>Email:</label>
+						<input type='email' name='email' id='email' />
+					</section>
+					<section className='input'>
+						<label htmlFor='password'>Password: </label>
+						<input type='password' name='password' id='password' />
+					</section>
+					<button>OK</button>
+				</section>
+			</div>
+			<div className='App-footer'>
+				<p>
+					Copyright {getFullYear()} - {getFooterCopy(true)}
+				</p>
+			</div>
+		</div>
+	);
 }
 
 export default App;
